refactor: wsclient, better error handling, channel based subscriptions
