+++ main.go
@@ -7,6 +7,9 @@ import (
 )
 
 func main() {
+	var stopCh chan struct{}
+	var subUnsubCh chan wsclient.SubUnsubRequest
+
 	ws := wsclient.New()
 
 	wsSubUnsubEventHandler := func(event *wsclient.SubUnsubEvent) {
@@ -15,13 +18,23 @@ func main() {
 
 	wsTradeEventHandler := func(event *wsclient.TradeEvent) {
 		log.Printf("Trade event: %+v\n", event)
+
+		// example
+		subUnsubCh <- wsclient.SubUnsubTradeTopics(wsclient.SUBSCRIBE,[]string{"btcusdc", "ethusdc"})
 	}
 
 	wsErrHandler := func(err error) {
-		log.Printf("Error: %+v\n", err)
+		switch err {
+		case err.(*wsclient.ErrSubUnsub):
+			log.Println("SubUnsub error: ", err)
+			stopCh <- struct{}{}
+
+		default:
+			log.Println("Default: ", err)
+		}
 	}
 
-	doneCh, stopCh, err := ws.Trade([]string{"btcusdc", "ethusdc"}, wsTradeEventHandler, wsSubUnsubEventHandler, wsErrHandler)
+	doneCh, stopCh, subUnsubCh, err := ws.Trade([]string{"btcusdc", "ethusdc"}, wsTradeEventHandler, wsSubUnsubEventHandler, wsErrHandler)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -34,3 +47,4 @@ func main() {
 		log.Println("done was called")
 	}
 }
+
